Creazione Post

+++ app/src/main/java/com/example/progettosocial/ui/HomeFragment.java
@@ -15,6 +15,7 @@ import android.view.ViewGroup;
 import android.widget.Toast;
 
 import com.example.progettosocial.api.ApiManager;
+import com.example.progettosocial.api.dto.request.CreatePostRequest;
 import com.example.progettosocial.api.dto.response.LastPostResponse;
 import com.example.progettosocial.api.dto.response.PostDTO;
 import com.example.progettosocial.api.dto.response.UtenteInfoDTO;
@@ -40,6 +41,8 @@ public class HomeFragment extends Fragment implements Callback {
     FragmentHomeBinding binding;
     DBManager db;
     PostDAO postDao;
+    PostAdapter adapter;
+
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -73,7 +76,7 @@ public class HomeFragment extends Fragment implements Callback {
             return insets;
         });
         binding.recyclerViewPosts.setLayoutManager(new LinearLayoutManager(getContext()));
-        PostAdapter adapter = new PostAdapter(postDao.getAll());
+        adapter = new PostAdapter(postDao.getAll());
         binding.recyclerViewPosts.setAdapter(adapter);
 
         binding.PostBox.setEndIconOnClickListener(v -> {
@@ -81,16 +84,9 @@ public class HomeFragment extends Fragment implements Callback {
             if(contenutoPost.isEmpty()) {
                 Toast.makeText(getContext(), "Post Vuoto", Toast.LENGTH_SHORT).show();
             }else {
-                /*
-                Post nuovoPost = new Post(null, user.getUsername(), contenutoPost, LocalDateTime.now().format(formatter).toString(), null, null, false, true);
-                postDao.insertPost(nuovoPost);
-                PostAdapter adapterNuovo = new PostAdapter(postDao.getAll());
-                binding.recyclerViewPosts.setAdapter(adapterNuovo);
-
-
+                CreatePostRequest nuovoPostRequest = new CreatePostRequest(contenutoPost);
+                ApiManager.getInstance().creaPost(nuovoPostRequest, this, requireContext());
                 binding.PostContent.setText(null);//svuota il campo dopo avere creato il post
-
-                 */
             }
         });
 
@@ -105,28 +101,45 @@ public class HomeFragment extends Fragment implements Callback {
 
     @Override
     public void onResponse(@NonNull Call call, @NonNull Response response) throws IOException {
-        if (!response.isSuccessful()){
-            requireActivity().runOnUiThread(()->{
-                Toast.makeText(requireContext(), "Impossibile caricare i post", Toast.LENGTH_SHORT).show();
-            });
-        }
-        else{
-            ObjectMapper mapper = new ObjectMapper();
-            try {
-                LastPostResponse lastPost = mapper.readValue(response.body().string(), LastPostResponse.class);
-                requireActivity().runOnUiThread(()->{
-                    db = DBManager.getInstance(requireContext());
-                    postDao = db.getPostDao();
-                    postDao.deleteAll();
-                    for(PostDTO post : lastPost.getPostList()){
-                        postDao.insertPost(new Post(post.getId(),post.getAutore().getUsername(), post.getTesto(), post.getDataPubblicazione(), post.getNumeroLikes(), post.getNumeroCommenti(), post.isLiked(), post.isMine()));
-                    }
+        String url = call.request().url().toString();
+        if (url.contains("getLastPost")) {
+
+            if (!response.isSuccessful()) {
+                requireActivity().runOnUiThread(() -> {
+                    Toast.makeText(requireContext(), "Impossibile caricare i post", Toast.LENGTH_SHORT).show();
                 });
-            }
-            catch(Exception e){
-                throw new RuntimeException(e);
+            } else {
+                ObjectMapper mapper = new ObjectMapper();
+                try {
+                    LastPostResponse lastPost = mapper.readValue(response.body().string(), LastPostResponse.class);
+                    requireActivity().runOnUiThread(() -> {
+                        db = DBManager.getInstance(requireContext());
+                        postDao = db.getPostDao();
+                        postDao.deleteAll();
+                        for (PostDTO post : lastPost.getPostList()) {
+                            postDao.insertPost(new Post(post.getId(), post.getAutore().getUsername(), post.getTesto(), post.getDataPubblicazione(), post.getNumeroLikes(), post.getNumeroCommenti(), post.isLiked(), post.isMine()));
+                        }
+                        binding.recyclerViewPosts.setLayoutManager(new LinearLayoutManager(getContext()));
+                        adapter = new PostAdapter(postDao.getAll());
+                        binding.recyclerViewPosts.setAdapter(adapter);
+                    });
+                } catch (Exception e) {
+                    throw new RuntimeException(e);
+                }
             }
 
+        }else if(url.contains("createPost")){
+            if (response.isSuccessful()) {
+                requireActivity().runOnUiThread(() -> {
+                    ApiManager.getInstance().getLastPost(this, requireContext());
+                });
+            } else {
+                String body=response.body().string();
+                requireActivity().runOnUiThread(()->{
+                    Toast.makeText(getContext(),body, Toast.LENGTH_LONG).show();
+                });
+            }
         }
+
     }
 }
